extract reader
